# Remove webp dependency and some other performance tweaks