--- a/imagemeta.go
+++ b/imagemeta.go
@@ -4,18 +4,21 @@ import (
 	"encoding/binary"
 	"fmt"
 	"io"
-	"math/big"
-	"time"
 )
 
 const (
-	TagSourceEXIF TagSource = iota
+	TagSourceEXIF TagSource = 1 << iota
 	TagSourceIPTC
 	TagSourceXMP
 )
 
-// Sentinel error to signal that the walk should stop.
-var ErrStopWalking = fmt.Errorf("stop walking")
+var (
+	// Sentinel error to signal that the walk should stop.
+	ErrStopWalking = fmt.Errorf("stop walking")
+
+	// Internal error to signal that we should stop any further processing.
+	errStop = fmt.Errorf("stop")
+)
 
 const (
 	ImageFormatAuto ImageFormat = iota
@@ -35,8 +38,8 @@ func Decode(opts Options) (err error) {
 	if opts.ImageFormat == ImageFormatAuto {
 		return fmt.Errorf("need an image format; format detection not implemented yet")
 	}
-	if opts.SourceSet == nil {
-		opts.SourceSet = map[TagSource]bool{TagSourceEXIF: true, TagSourceIPTC: true, TagSourceXMP: true}
+	if opts.Sources == 0 {
+		opts.Sources = TagSourceEXIF | TagSourceIPTC | TagSourceXMP
 	}
 
 	br := &streamReader{
@@ -66,6 +69,7 @@ func Decode(opts Options) (err error) {
 	case ImageFormatJPEG:
 		dec = &imageDecoderJPEG{baseStreamingDecoder: base}
 	case ImageFormatWebP:
+		base.byteOrder = binary.LittleEndian
 		dec = &decoderWebP{baseStreamingDecoder: base}
 	case ImageFormatPNG:
 		dec = &imageDecoderPNG{baseStreamingDecoder: base}
@@ -78,7 +82,10 @@ func Decode(opts Options) (err error) {
 	if err == ErrStopWalking {
 		return nil
 	}
-	return err
+	if err != nil {
+		return err
+	}
+	return base.err
 }
 
 type HandleTagFunc func(info TagInfo) error
@@ -86,92 +93,28 @@ type HandleTagFunc func(info TagInfo) error
 //go:generate stringer -type=ImageFormat
 type ImageFormat int
 
-// Meta contains the EXIF and IPTC metadata.
-// TODO1
-type Meta struct {
-	EXIF Tags
-	IPTC Tags
-}
-
-// DateTime returns the DateTime tag as a time.Time parsed with the given location.
-// Todo timezone if loc is nil.
-func (m Meta) DateTime(loc *time.Location) time.Time {
-	tags := m.EXIF
-	if v, ok := tags["DateTime"]; ok {
-		t, err := time.ParseInLocation("2006:01:02 15:04:05", v.(string), loc)
-		if err == nil {
-			return t
-		}
-	}
-	return time.Time{}
-}
-
-// DateTimeOriginal returns the DateTimeOriginal tag as a time.Time parsed with the given location.
-func (m Meta) DateTimeOriginal(loc *time.Location) time.Time {
-	tags := m.EXIF
-	if v, ok := tags["DateTimeOriginal"]; ok {
-		t, err := time.ParseInLocation("2006:01:02 15:04:05", v.(string), loc)
-		if err == nil {
-			return t
-		}
-	}
-	return time.Time{}
-}
-
-func (m Meta) LatLong() (float64, float64) {
-	tags := m.EXIF
-	longv := tags.get("GPSLongitude")
-	ewv := tags.get("GPSLongitudeRef")
-	latv := tags.get("GPSLatitude")
-	nsv := tags.get("GPSLatitudeRef")
-
-	fmt.Printf("longv: %v, ewv: %v, latv: %v, nsv: %v\n", longv, ewv, latv, nsv)
-
-	if longv == nil || latv == nil {
-		return 0, 0
-	}
-
-	long := tags.toDegrees(longv)
-	lat := tags.toDegrees(latv)
-
-	if ewv == "W" {
-		long *= -1.0
-	}
-	if nsv == "S" {
-		lat *= -1.0
-	}
-
-	return lat, long
-}
-
-// Orientation returns the Orientation tag.
-func (m Meta) Orientation() Orientation {
-	tags := m.EXIF
-	if v, ok := tags["Orientation"]; ok {
-		return Orientation(v.(uint16))
-	}
-	return OrientationUnspecified
-}
-
 type Options struct {
 	// The Reader (typically a *os.File) to read image metadata from.
 	R Reader
 
+	// The image format in R.
 	ImageFormat ImageFormat
 
+	// The function to call for each tag.
 	HandleTag HandleTagFunc
 
-	// If set, the decoder will only read the given tag sources
-	SourceSet map[TagSource]bool
+	// If set, the decoder will only read the given tag sources.
+	// Note that this is a bitmask and you may send multiple sources at once.
+	Sources TagSource
 }
 
-type Orientation int
-
+// Reader is the interface that wraps the basic Read and Seek methods.
 type Reader interface {
 	io.ReadSeeker
 	io.ReaderAt
 }
 
+// TagInfo contains information about a tag.
 type TagInfo struct {
 	// The tag source.
 	Source TagSource
@@ -183,42 +126,12 @@ type TagInfo struct {
 	Value any
 }
 
+// TagSource is a bitmask and you may send multiple sources at once.
+//
 //go:generate stringer -type=TagSource
-type TagSource int
+type TagSource uint32
 
-// Tags is a map of EXIF and IPTC tags.
-type Tags map[string]any
-
-func (tags Tags) toDegrees(v any) float64 {
-	if v == nil {
-		return 0
-	}
-	f := [3]float64{}
-	for i, vv := range v.([]interface{}) {
-		if i >= 3 {
-			break
-		}
-		r := vv.(*big.Rat)
-		f[i], _ = r.Float64()
-	}
-	// TODO1 other types?
-	return f[0] + f[1]/60 + f[2]/3600.0
-}
-
-func (tags Tags) get(key string) any {
-	if v, ok := tags[key]; ok {
-		return v
-	}
-	return nil
-}
-
-type closerFunc func() error
-
-func (f closerFunc) Close() error {
-	return f()
-}
-
-type readerCloser interface {
-	Reader
-	io.Closer
+// Has returns true if the given source is set.
+func (t TagSource) Has(source TagSource) bool {
+	return t&source != 0
 }