--- a/tagsource_string.go
+++ b/tagsource_string.go
@@ -4,22 +4,32 @@ package imagemeta
 
 import "strconv"
 
-const _TagSource_name = "TagSourceEXIFTagSourceIPTCTagSourceXMP"
-
-var _TagSource_index = [...]uint8{0, 13, 26, 38}
-
-func (i TagSource) String() string {
-	if i < 0 || i >= TagSource(len(_TagSource_index)-1) {
-		return "TagSource(" + strconv.FormatInt(int64(i), 10) + ")"
-	}
-	return _TagSource_name[_TagSource_index[i]:_TagSource_index[i+1]]
-}
-
 func _() {
 	// An "invalid array index" compiler error signifies that the constant values have changed.
 	// Re-run the stringer command to generate them again.
 	var x [1]struct{}
-	_ = x[TagSourceEXIF-0]
-	_ = x[TagSourceIPTC-1]
-	_ = x[TagSourceXMP-2]
+	_ = x[TagSourceEXIF-1]
+	_ = x[TagSourceIPTC-2]
+	_ = x[TagSourceXMP-4]
+}
+
+const (
+	_TagSource_name_0 = "TagSourceEXIFTagSourceIPTC"
+	_TagSource_name_1 = "TagSourceXMP"
+)
+
+var (
+	_TagSource_index_0 = [...]uint8{0, 13, 26}
+)
+
+func (i TagSource) String() string {
+	switch {
+	case 1 <= i && i <= 2:
+		i -= 1
+		return _TagSource_name_0[_TagSource_index_0[i]:_TagSource_index_0[i+1]]
+	case i == 4:
+		return _TagSource_name_1
+	default:
+		return "TagSource(" + strconv.FormatInt(int64(i), 10) + ")"
+	}
 }