```
name                              old time/op    new time/op    delta
Decode/bep/imagemeta/jpg/exif-10    24.5µs ± 1%    17.8µs ± 4%  -27.46%  (p=0.029 n=4+4)

name                              old alloc/op   new alloc/op   delta
Decode/bep/imagemeta/jpg/exif-10    7.77kB ± 0%    6.01kB ± 0%  -22.62%  (p=0.029 n=4+4)

name                              old allocs/op  new allocs/op  delta
Decode/bep/imagemeta/jpg/exif-10       837 ± 0%       242 ± 0%  -71.09%  (p=0.029 n=4+4)
```